+++ multisig/proposals/unichainmainnet/vip-009/index.ts
@@ -2,18 +2,18 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 
 import { makeProposal } from "../../../../src/utils";
 
-const { unichainsepolia } = NETWORK_ADDRESSES;
+const { unichainmainnet } = NETWORK_ADDRESSES;
 
-export const ACM = "0x854C064EA6b503A97980F481FA3B7279012fdeDd";
+export const ACM = "0x1f12014c497a9d905155eB9BfDD9FaC6885e61d0";
 const DEFAULT_ADMIN_ROLE = "0x0000000000000000000000000000000000000000000000000000000000000000";
 
-export const vip008 = () => {
+export const vip009 = () => {
   return makeProposal([
     {
       target: ACM,
       signature: "grantRole(bytes32,address)",
-      params: [DEFAULT_ADMIN_ROLE, unichainsepolia.NORMAL_TIMELOCK],
+      params: [DEFAULT_ADMIN_ROLE, unichainmainnet.NORMAL_TIMELOCK],
     },
   ]);
 };
-export default vip008;
+export default vip009;
