+++ simulations/vip-456/bscmainnet.ts
@@ -5,12 +5,12 @@ import { LzChainId } from "src/types";
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import vip455, { MAX_DAILY_LIMIT, OMNICHAIN_PROPOSAL_SENDER } from "../../vips/vip-455/bsctestnet";
+import vip456, { MAX_DAILY_LIMIT, OMNICHAIN_PROPOSAL_SENDER } from "../../vips/vip-456/bscmainnet";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager_ABI.json";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
-const { unichainsepolia } = NETWORK_ADDRESSES;
-forking(48215809, async () => {
+const { unichainmainnet } = NETWORK_ADDRESSES;
+forking(46771559, async () => {
   const provider = ethers.provider;
   const omnichainProposalSender = new ethers.Contract(
     OMNICHAIN_PROPOSAL_SENDER,
@@ -20,14 +20,14 @@ forking(48215809, async () => {
 
   describe("Pre-VIP behaviour", () => {
     it("Daily limit should be 0", async () => {
-      expect(await omnichainProposalSender.chainIdToMaxDailyLimit(LzChainId.unichainsepolia)).to.equals(0);
+      expect(await omnichainProposalSender.chainIdToMaxDailyLimit(LzChainId.unichainmainnet)).to.equals(0);
     });
     it("Trusted remote should not be set", async () => {
-      expect(await omnichainProposalSender.trustedRemoteLookup(LzChainId.unichainsepolia)).to.be.equals("0x");
+      expect(await omnichainProposalSender.trustedRemoteLookup(LzChainId.unichainmainnet)).to.be.equals("0x");
     });
   });
 
-  testVip("vip455 give permissions to timelock", await vip455(), {
+  testVip("vip456 give permissions to timelock", await vip456(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
@@ -39,17 +39,17 @@ forking(48215809, async () => {
   });
 
   describe("Post-VIP behavior", () => {
-    it("Daily limit should be 100 of unichainsepolia", async () => {
-      expect(await omnichainProposalSender.chainIdToMaxDailyLimit(LzChainId.unichainsepolia)).to.equals(
+    it("Daily limit should be 100 of unichainmainnet", async () => {
+      expect(await omnichainProposalSender.chainIdToMaxDailyLimit(LzChainId.unichainmainnet)).to.equals(
         MAX_DAILY_LIMIT,
       );
     });
 
-    it("Trusted remote should be set of unichainsepolia", async () => {
-      expect(await omnichainProposalSender.trustedRemoteLookup(LzChainId.unichainsepolia)).to.be.equals(
+    it("Trusted remote should be set of unichainmainnet", async () => {
+      expect(await omnichainProposalSender.trustedRemoteLookup(LzChainId.unichainmainnet)).to.be.equals(
         ethers.utils.solidityPack(
           ["address", "address"],
-          [unichainsepolia.OMNICHAIN_GOVERNANCE_EXECUTOR, OMNICHAIN_PROPOSAL_SENDER],
+          [unichainmainnet.OMNICHAIN_GOVERNANCE_EXECUTOR, OMNICHAIN_PROPOSAL_SENDER],
         ),
       );
     });
